Internationalized Text
A white paper on internationalized text has been produced. This is an integration of Howards work from the Best Practice with an analysis of the support provided for internationalized text in the TC211 suite of standards. The objective of this white paper is to inform OGC policy on internationalized text.  It may also be included as an annex to the POI Standard.